% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_identical.R
\name{all_identical}
\alias{all_identical}
\title{All Elements Of A List Are Identical}
\usage{
all_identical(l)
}
\arguments{
\item{l}{a list of elements}
}
\value{
a boolean value
}
\description{
Check if all elements of a list are identical. This means both that the elements are the same in content, and the same in order.
}
\examples{
\dontrun{
if(interactive()){
 test_F = replicate(5, sample(colnames(mtcars)), simplify = F)
 test_T = rep(colnames(mtcars), 5, simplify = F)
 all_identical(l = test_F)
 all_identical(l = test_T)
 }
}
}
