% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{build_coordinates}
\alias{build_coordinates}
\title{FUNCTION_TITLE}
\usage{
build_coordinates(mat, method = "tsne", ...)
}
\arguments{
\item{mat}{input matrix (each row is a variable / a gene and each column is an observation / a cell.}

\item{method}{character string defining the dimension-reduction method used. One of 'umap' or 'tsne', calling umap::umap or Rtsne::Rtsne, respectively. Default: 'tsne'}

\item{...}{other arguments passed to dimension-reduction method (umap or Rtsne)}
}
\value{
coordinates embedding matrix
}
\description{
FUNCTION_DESCRIPTION
}
