% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{build_knn}
\alias{build_knn}
\title{Search Nearest Neighbours}
\usage{
build_knn(mat, k = seq(5, 30, 1), algorithm = "kd_tree")
}
\arguments{
\item{mat}{coordinates embedding matrix (such as that generated with build_tsne() or build_umap())}

\item{k}{a numeric value or vector defining the maximum number or numbers of nearest neighours to search. If a vector is provided, the function will iterate for every value of k. Default: seq(5, 30, 1)}

\item{algorithm}{nearest neighbour searching algorithm. One of kd_tree, cover_tree, CR, brute. Default: 'kd_tree'}
}
\value{
a dataframe with the k-nearest neighbours and corresponding weights (1/1 + the euclidean distance).
}
\description{
computes fast k-nearest neighbour search. Wrapper for FNN::get.knn().
}
\seealso{
\code{\link[FNN]{get.knn}}
}
