% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_coord.R
\name{build_tsne}
\alias{build_tsne}
\title{t-Distributed Stochastic Neighbour Embedding (Barnes-Hut implementation)}
\usage{
build_tsne(mat, ...)
}
\arguments{
\item{mat}{input matrix (each row is a variable / a gene and each column is an observation / a cell.}

\item{...}{other arguments passed to Rtsne}
}
\value{
coordinates embedding matrix
}
\description{
Computes a low dimensional embedding of high-dimensional data and returns the embedding coordinates matrix. Wrapper function for Rtsne::Rtsne().
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
