% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{build_umap}
\alias{build_umap}
\title{Uniform Manifold Approximation and Projection}
\usage{
build_umap(mat, dist.method = "pearson", spread = 0.5,
  min.dist = 0.1, ...)
}
\arguments{
\item{mat}{input matrix (each row is a variable / a gene and each column is an observation / a cell.}

\item{dist.method}{a character string for the distance metric used. One of pearson, euclidean, manhattan, cosine or pearson2. Default: 'pearson'}

\item{spread}{a numeric value that is used during automatic estimation of a/b parameters. Default: 0.5}

\item{min.dist}{a numeric value that determines how close points appear in the final layout. Default: 0.1}

\item{...}{other arguments passed to umap}
}
\value{
coordinates embedding matrix
}
\description{
Computes a manifold approximation and projection and returns the embedding coordinates matrix. Wrapper function for umap::umap().
}
\seealso{
\code{\link[umap]{umap.defaults}},\code{\link[umap]{c("umap", "umap")}}
}
