% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_by_genes.R
\name{clean_by_genes}
\alias{clean_by_genes}
\title{clean_by_genes}
\usage{
clean_by_genes(L, nsig1 = 30, nsig2 = 5)
}
\arguments{
\item{L}{List of clusters' differentially expressed genes}

\item{nsig1}{cutoff for number of differentially expressed genes. Default: 30}

\item{nsig2}{cutoff for number of most differentially expressed genes. Default: 5}
}
\value{
a filtered list of statistically significant clusters. ie. those passing nsig1 and nsig2 cutoffs.
}
\description{
Filter for Statistically Significant Clusters
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
