% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_analysis.R
\name{cluster_graph}
\alias{cluster_graph}
\title{Cluster igraph graphs}
\usage{
cluster_graph(graph, method = "louvain")
}
\arguments{
\item{graph}{igraph graph object.}

\item{method}{a character string, one of 'louvain' or 'hierarchical', defining the clustering method used. Default: 'louvain'}
}
\value{
a numeric vector of member IDs whose names are the members (or member IDs), or a list of numeric vectors if a list of igraph graphs was provided.
}
\description{
Compute and retrieve clusters from igraph graphs.
}
\seealso{
\code{\link[igraph]{cluster_louvain}},\code{\link[igraph]{cluster_walktrap}},\code{\link[igraph]{membership}}
}
