% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusters}
\alias{clusters}
\title{clusters}
\usage{
clusters(m = NULL, hc = NULL, cor.method = "pearson",
  hc.method = "average", compute.dist = T, dist.method = "euclidean",
  k = NULL, h = NULL, clean = TRUE, min.rel = 0.05,
  max.rel = 0.8, min.abs = 5)
}
\arguments{
\item{m}{matrix. Default: NULL}

\item{hc}{hierarchical clustering (hclust) object. If provided, will be used instead of computed. Default: NULL}

\item{cor.method}{a character string indicating which correlation coefficient is to be computed. Default: 'pearson'}

\item{hc.method}{a character string indicating which agglomeration method to use. Default: 'average'}

\item{compute.dist}{a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T}

\item{dist.method}{a character string specifying the distance metric to be used for hierarchical clustering. Default: 'euclidean'}

\item{k}{number of clusters to return. Default: NULL}

\item{h}{tree height at which to retrieve clusters. Default: NULL}

\item{clean}{if TRUE, will return clusters that are within the sizes specified. Default: TRUE}

\item{min.rel}{minimum relative size of a cluster. Default: 0.01 (ie. 0.01 of total number of variables)}

\item{max.rel}{maximum relative size of a cluster. Default: 0.8 (ie. 0.8 of total number of variables)}

\item{min.abs}{absolute minimum size. If min.rel < min.abs, min.abs will replace min.rel. Default: 5}
}
\value{
with default parameters, will compute hierarchical clustering object of the pariwise correlations of the matrix provided and retrieve all clusters at all possible heights. If the user prefers a specified number of clusters or a specified height, these can instead be provided.
}
\description{
Compute and Retrieve Clusters from Hierarchical Clustering at all Heights
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{cutree}}
}
