% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{FUNCTION_TITLE}
\usage{
corPlot(cormat, lim.find = T, limits = c(-0.2, 0.2),
  x.name = "Cells", y.name = "Cells", title = NULL,
  subtitle = NULL, cols = heatCols, ...)
}
\arguments{
\item{cormat}{PARAM_DESCRIPTION}

\item{lim.find}{PARAM_DESCRIPTION, Default: T}

\item{limits}{PARAM_DESCRIPTION, Default: c(-0.2, 0.2)}

\item{x.name}{PARAM_DESCRIPTION, Default: 'Cells'}

\item{y.name}{PARAM_DESCRIPTION, Default: 'Cells'}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{cols}{PARAM_DESCRIPTION, Default: heatCols}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{melt}}
}
