% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_detected.R
\name{count_detected}
\alias{count_detected}
\title{FUNCTION_TITLE}
\usage{
count_detected(m)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[matrixStats]{rowCounts}}
 \code{\link[stats]{setNames}}
}
