% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creorder.R
\name{creorder}
\alias{creorder}
\title{Generate ordered correlation matrices on the fly, FLexiblY.}
\usage{
creorder(mat = NULL, hc.method = "average", cor.method = "pearson",
  compute.dist = T, dist.method = "euclidean", ord.labels = T)
}
\arguments{
\item{mat}{matrix. Default: NULL}

\item{hc.method}{a character string indicating which agglomeration method to use. Default: 'average'}

\item{cor.method}{a character string indicating which correlation coefficient is to be computed. Default: 'pearson'}

\item{compute.dist}{a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T}

\item{dist.method}{a character string specifying the distance metric to be used for hierarchical clustering. Default: ''euclidean'}

\item{ord.labels}{if FALSE, will return ordered indices rather than character vector. Default: T}
}
\value{
an ordered correlation matrix with the same number of rows and columns as there were columns in the original matrix.
}
\description{
Calls the scrabble function clusta() to generate a correlation matrix with the parameters specified and subsequently order the columns and rows by hierarchical clustering, again with parameters specified by the user (defaults are recommended for single-cell expression (RNAseq) data).
}
\seealso{
\code{\link[scrabble]{clusta}}
}
