% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{DE}
\alias{DE}
\title{DE}
\usage{
DE(m, m2 = NULL, group = NULL, is.log = T, FC = 2,
  adjust.method = "BH", p = 0.01, p.sort = F, return.all = F,
  return.p = F, fast = FALSE)
}
\arguments{
\item{m}{matrix 1}

\item{m2}{matrix 2. If null, m2 = m[, !group]. Defaut: NULL}

\item{group}{if m2 not supplied, m = m[, group]; m2 = m[, !group]. Default: NULL}

\item{is.log}{values are in log2 (used in fold_changes), Default: T}

\item{FC}{fold change cutoff, Default: 2}

\item{adjust.method}{Correction for multiple tests. If not desired, set to 'none'. Default: 'BH'}

\item{p}{p-value cutoff, Default: 0.01}

\item{p.sort}{returning values are sorted by p-value (most significant first), Default: F}

\item{return.all}{return all values (p-values and fold-change values), Default: F}

\item{return.p}{return p-values, Default: F}

\item{fast}{if TRUE, computes p-values for significant fold-change values only, Default: FALSE}
}
\value{
(log2) fold-change values for rows in m, sorted by FC (highest first) and including only those that are significant according to FC and p cutoffs.
}
\description{
Differential Expression between Groups
}
\details{
DETAILS
}
