% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea}
\alias{dea}
\title{Differential Expression Analysis}
\usage{
dea(x, y, is.log = T, fc = 2L, p = 0.05, use.wilcox = F,
  adjust.method = unique(c("BH", stats::p.adjust.methods)),
  sort.by = c("fc", "p"), output = c("fc", "p"))
}
\arguments{
\item{x}{matrix or character vector with matrix column names}

\item{y}{matrix to compare against or matrix with x and y column names}

\item{is.log}{logical indicating whether expression data is in log2 space. Fold change cutoff and calculation are changed accordingly. Default: T}

\item{fc}{fold change cutoff. Set to NULL to not filter genes based on fold change. Default: 2}

\item{p}{p value cutoff. Set to NULL to not filter genes based on p-values. Default: 0.05}

\item{use.wilcox}{logical value to use Wilcoxon-Mann-Whitney test instead of Student's t-test. Default: F}

\item{adjust.method}{correction for multiple testing? 'none' if not desired. Default: unique(c("BH", stats::p.adjust.methods))}

\item{sort.by}{sort output by 'fc' or 'p' value (best first). Set sort.by = NULL to preserve original gene (row) order. Default: 'fc'}

\item{output}{output 'fc' values or 'p' values (per gene). If output = NULL, a list containing both is returned. Default: 'fc'}
}
\value{
a numeric vector of gene fold changes or p-values or a list containing both. skip.genes = F to return information on every gene in the expression matrix (regardless of it passing cutoffs).
}
\description{
Differential Expression Analysis. Compute gene-level fold changes and (adjusted) p-values between cell clusters of an expression matrix.
}
\seealso{
\code{\link[stats]{character(0)}}
}
