% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Test data for scrabble}
\format{An object of class \code{list} of length 3.}
\usage{
demo_data
}
\description{
"demo_data" is a data list with the length of 3. The first element in
the list is generated drop-out scRNAseq data with 800 genes and 1000 cells. The second
element in the list is the generated bulk RNAseq data with 800 genes. The third
element is the true scRNAseq data without dropouts. The steps of generating the data
is shown in Details section.
}
\details{
The data set was generated from the well-developed R package Splatter.
}
\author{
Tao Peng, Kai Tan
}
\keyword{datasets}
