% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_changes.R
\name{fold_changes}
\alias{fold_changes}
\title{fold_changes}
\usage{
fold_changes(m, m2 = NULL, group = NULL, is.log = T)
}
\arguments{
\item{m}{matrix 1}

\item{m2}{matrix 2. If null, m2 = m[, !group]. Defaut: NULL}

\item{group}{if m2 not supplied, m = m[, group]; m2 = m[, !group]. Default: NULL}

\item{is.log}{values are in log2 (used in fold_changes), Default: T}
}
\value{
fold change values for rows in m
}
\description{
Compute (log2) Fold-Changes between Groups
}
\details{
DETAILS
}
