% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{genes}
\alias{genes}
\title{genes}
\usage{
genes(m, hc = NULL, clusters = NULL, FC = 2, is.log = T,
  p = 0.01, adjust.method = "BH", p.sort = F, return.all = F,
  return.p = F, fast = T, clean = TRUE, nsig1 = 30, nsig2 = 5,
  Names = F, return.clusters = F, ...)
}
\arguments{
\item{m}{matrix}

\item{hc}{hierarchical clustering (hclust) object. If provided, will be used instead of computed. Default: NULL}

\item{clusters}{List of cluster memberships (character vectors). If provided, will be used instead of computed.. Default: NULL}

\item{FC}{fold change cutoff. Default: 2}

\item{is.log}{values are in log2 (used in fold_changes). Default: T}

\item{p}{p-value cutoff. Default: 0.01}

\item{adjust.method}{Correction for multiple tests. If not desired, set to 'none'. Default: 'BH'}

\item{p.sort}{returning values are sorted by p-value (most significant first). Default: F}

\item{return.all}{return cluster membership and cluster DE genes. Default: F}

\item{return.p}{return cluster DE p-values. Default: F}

\item{fast}{if TRUE, computes p-values for significant fold-change values only. Default: T}

\item{clean}{return only significant clusters as defined by nsig1 AND nsig2. Default: TRUE}

\item{nsig1}{cutoff for number of differentially expressed genes. Default: 30}

\item{nsig2}{cutoff for number of most differentially expressed genes. Default: 5}

\item{Names}{if TRUE, will return only gene names. Default: F}

\item{return.clusters}{if TRUE, will return cluster memberships as well as cluster genes. Default: F}

\item{...}{other arguments passed to \code{clusters}.}
}
\value{
with default parameters, will return a list of differentially expressed genes in statistically significant clusters. Differentially expressed genes are those passing fold change and p-value cutoffs. Statistically significant clusters are those passing nsig1 and nsig2 cutoffs. Within clusters, gene values displayed are fold changes and genes are ordered by highest FC first.
}
\description{
Differentially Expressed Genes of Statistically Significant Clusters
}
\details{
DETAILS
}
