% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{Get words from random letters}
\usage{
get_words(letters, word_len = 0, wildcard = FALSE)
}
\arguments{
\item{letters}{random string of letters}

\item{word_len}{length of desired word}

\item{wildcard}{boolean indicating if there's a wildcard}
}
\value{
returns a list with words made up from string of letters
}
\description{
Get words from random letters
}
\examples{
get_words("uhwoih")
get_words("ihfjwd", 4)
get_words("hasfic", 4, TRUE)
}
