% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC.R
\name{HC}
\alias{HC}
\title{HC}
\usage{
HC(m = NULL, cr = FALSE, hc = FALSE, ord = FALSE,
  clusters = FALSE, returnSteps = FALSE, hc.method = "average",
  cor.method = "pearson", compute.dist = T,
  dist.method = "euclidean", ord.labels = T, h = NULL, k = NULL,
  min.cluster.size = 5, max.cluster.size = 0.8)
}
\arguments{
\item{m}{matrix. Default: NULL}

\item{cr}{correlation matrix. If provided, cr will not be computed with m. Default: FALSE}

\item{hc}{hclust object. If provided, hc will not be computed with cr. Default: FALSE}

\item{ord}{ordered character vector, retrieved from hc. Default: FALSE}

\item{clusters}{list of clusters (character vectors), retrieved from hc. Default: FALSE}

\item{hc.method}{a character string indicating which agglomeration method to use. Default: 'average'}

\item{cor.method}{a character string indicating which correlation coefficient is to be computed. Default: 'pearson'}

\item{compute.dist}{a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T}

\item{dist.method}{a character string specifying the distance metric to be used for hierarchical clustering. Default: ''euclidean'}

\item{ord.labels}{if FALSE, will return ordered indices rather than character vector. Default: T}
}
\value{
The function proceeds through the following steps: (1) compute correlation matrix, (2) compute hierarchical clustering object, (3) retrieve ordered character vector. It will start at the point for which an object is provided. i.e. if cr is provided, then step (1) is skipped. It will either return a list with all computed objects \code{list(cr = cr, hc = hc, ord = ord)} or stop at the point at which an argument is set to TRUE. i.e. if cr = T, step (3) is skipped and
only the correlation matrix is returned. Thus, users can provide an expression matrix and ask for the ordered columns' character vector, or provide a correlation matrix and ask for the corresponding hierarchical clustering object and so on.
}
\description{
Hierarchical clustering
}
\seealso{
\code{\link[stats]{cor}},\code{\link[stats]{hclust}},\code{\link[stats]{dist}}
}
