% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_word.R
\name{is_word}
\alias{is_word}
\title{Is Word}
\usage{
is_word(word)
}
\arguments{
\item{word}{string of letter}
}
\value{
string saying if the input is an actual word or not
}
\description{
Is Word
}
\examples{
is_word("hello")
is_word("hbkjwfen")
}
