% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_filter.R
\name{jaccard_filter}
\alias{jaccard_filter}
\title{Ordered List Filter by Jaccard Similarity}
\usage{
jaccard_filter(L, cutoff = 0.8, return.index = FALSE)
}
\arguments{
\item{L}{List of character vectors}

\item{cutoff}{numeric value between 0 and 1. Default: 0.8}

\item{return.index}{boolean value indicating whether a vector of indexes to keep should be returned instead of the filtered list. Default: FALSE}
}
\value{
Filtered list of character vectors whose pairwise Jaccard similarities are all smaller than or equal to the cutoff specified.
}
\description{
Computes the jaccard similarities of all pairs of vectors, and, for any pair with too-high similarity, removes the vector that is placed lower in the list.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
