% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logTPM}
\alias{logTPM}
\title{FUNCTION_TITLE}
\usage{
logTPM(m, dividebyten = TRUE)
}
\arguments{
\item{m}{PARAM_DESCRIPTION}

\item{dividebyten}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
