% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_combinations.R
\name{map_combinations}
\alias{map_combinations}
\title{Map a Function to All Combinations of Vectors}
\usage{
map_combinations(x, y = NULL, FUN = jaccard)
}
\arguments{
\item{x}{a vector or a list of vectors}

\item{y}{a vector or a list of vectors. If y is NULL, y will be set to x, such that if x is a list, the function FUN will be applied to all pairs of vectors in the list. Default: NULL}

\item{FUN}{the function to apply to all pairs of x and y. Default: jaccard}
}
\value{
a matrix of resulting values after applying FUN to all pairs of x and y (or the one pair if x and y are both atomic vectors). The number of columns and the number of rows will correspond to the lengths of x and y respectively.
}
\description{
F
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{setNames}}
}
