% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{Binary "in or out" Matrix}
\usage{
membership(X, universe = NULL, order = F, col.order = order,
  row.order = order, dist.method = "manhattan")
}
\arguments{
\item{X}{a list or matrix of character vectors or columns}

\item{order}{a boolean value indicating whether the resulting membership matrix should have rows and columns reordered. Default: F}

\item{col.order}{a boolean value indicating whether the resulting membership matrix should have columns reordered. Default: F}

\item{row.order}{a boolean value indicating whether the resulting membership matrix should have rows reordered. Default: F}

\item{dist.method}{a character string indicating the distance method to be used for the hierarchichal ordering. e.g. 'manhattan', 'euclidean'. Default: 'manhattan'}
}
\value{
a matrix of 0s and 1s for each of the unique strings in X in each of the vectors or columns in X.
}
\description{
Computes the membership of strings in a series of character vectors, provided either in list or matrix format (vector == column).
}
