% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{FUNCTION_TITLE}
\usage{
modes(m = NULL, X = NULL, prob = 0.95, coverage = 0.8, size = 10,
  exclude.unassigned = T)
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{X}{PARAM_DESCRIPTION, Default: NULL}

\item{prob}{PARAM_DESCRIPTION, Default: 0.95}

\item{coverage}{PARAM_DESCRIPTION, Default: 0.8}

\item{size}{PARAM_DESCRIPTION, Default: 10}

\item{exclude.unassigned}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[mixtools]{normalmixEM}}
}
