% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mots_scrabble.R
\docType{data}
\name{mots_scrabble}
\alias{mots_scrabble}
\title{Dictionnaire de mots en français pour le scrabble}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 372291 rows and 2 columns.}
\usage{
mots_scrabble
}
\description{
Dictionnaire de mots en français pour le scrabble
}
\keyword{datasets}
