% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_words.R
\name{num_words}
\alias{num_words}
\title{Number of words}
\usage{
num_words(letters, word_len = 0, wildcard = FALSE)
}
\arguments{
\item{letters}{string of characters to build words}

\item{word_len}{if desired, length of words to be returned}

\item{wildcard}{boolean to activate or not a wildcard}
}
\value{
returns an int with the number of possible words you can create
}
\description{
Number of words
}
\examples{
get_words("uhwoih")
get_words("ihfjwd", word_len = 4)
get_words("hasfic", wildcard = TRUE)
}
