% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by_significance.R
\name{order_by_significance}
\alias{order_by_significance}
\title{Order List of Named Gene Vectors by Significance}
\usage{
order_by_significance(L, method = c("combined", "sequential", "sig1",
  "sig2"), ties.method = c("average", "first", "last", "random", "max",
  "min"), first = c("sig1", "sig2"), highest = TRUE,
  decreasing = TRUE, return.index = FALSE)
}
\arguments{
\item{L}{List of named numeric vectors (e.g. Gene names and FC values for different clusters)}

\item{method}{character string indicating which ordering method should be used, one of 'combined', 'sequential', 'sig1' or 'sig2'. 'combined' orders by the mean of sig1 ranks and sig2 ranks. 'sequential' orders by sig1 and sig2 sequentially, ordering by the second only where there are ties in the first. 'sig1' orders by the number of elements in each vector. 'sig2' orders by the number of elements in each vector after filtering the list such that each element only appears once in the vector in which it's value is highest. Default: 'combined'}

\item{ties.method}{character string specifying how ties are treated, see base::rank for more details.}

\item{first}{character string indicating whether ordering should be first by sig1 and then by sig2 (default) or vice versa. One of 'sig1' or 'sig2'. Only relevant if method is 'sequential'. Default: 'sig1'}

\item{decreasing}{boolean value indicating whether the list should be reordered by decreasing or increasing significance. If TRUE, the most significant vector will appear first. Default: TRUE}

\item{return.index}{boolean value indicating whether ordering index vector should be returned instead of sorted list. Default: FALSE}
}
\value{
List of named gene vectors reordered by significance.
}
\description{
Reorders a list of named numeric vectors according to two parameters for significance, sig1 and sig2. sig1 is the number of elements in each vector. sig2 is the number of elements in each vector after filtering the list such that each element only appears once in the vector in which it's value was highest.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
