% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_data}
\alias{plot_data}
\title{It is used to plot the data as a images}
\usage{
plot_data(data, name)
}
\arguments{
\item{data}{the input data which is a data matrix.}

\item{name}{the name of the data and it is a string.}
}
\value{
ggplot handler
}
\description{
It is used to plot the data as a images
}
\examples{
# Run plot_data
p <- plot_data(demo_data[[1]],"Drop-out Data")

}
