% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{Copmutes the number of points in a scrabble word.}
\usage{
points(word, point.mult = rep(1, nchar(word)), dw = 0, tw = 0)
}
\arguments{
\item{word}{# string of letters}

\item{dw}{# number of double word scores}

\item{tw}{# number of tripple word scores}

\item{locations}{# list of special letter score tile locations within the word}

\item{multiples}{# double or tripple letter score corrosponding to letter index}
}
\value{
# total score
}
\description{
Copmutes the number of points in a scrabble word.
}
\examples{
points("botch", c(1,2,1,2,1), dw = 2)
}
