% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixBasics.R
\name{range.rowMeans}
\alias{range.rowMeans}
\title{Range of the rowMeans of a matrix}
\usage{
\method{range}{rowMeans}(mat)
}
\arguments{
\item{a}{matrix}
}
\value{
the range of the rowMeans of the matrix provided.
}
\description{
Returns the range over the rowMeans of a matrix
}
