% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder a Matrix}
\usage{
reorder(mat = NULL, sim.mat = NULL, corr = TRUE,
  hc.method = "average", cor.method = "pearson", compute.dist = T,
  dist.method = "euclidean", both = T, row = both, col = both,
  ord.labels = T)
}
\arguments{
\item{mat}{matrix. Default: NULL}

\item{hc.method}{a character string indicating which agglomeration method to use. Default: 'average'}

\item{cor.method}{a character string indicating which correlation coefficient is to be computed. Default: 'pearson'}

\item{compute.dist}{a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T}

\item{dist.method}{a character string specifying the distance metric to be used for hierarchical clustering. Default: ''euclidean'}

\item{both}{a boolean value indicating whether both rows and columns should be reordered. Default: F}

\item{row}{a boolean value indicating whether rows should be reordered. Default: F}

\item{col}{a boolean value indicating whether columns should be reordered. Default: T}

\item{ord.labels}{if FALSE, will return ordered indices rather than character vector. Default: T}
}
\value{
the matrix, with rows and/or columns reordered.
}
\description{
Calls the scrabble function clusta() to order the columns and/or the rows of a matrix by hierarchical clustering.
}
\seealso{
\code{\link[scrabble]{clusta}}
}
