% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scLDA}
\alias{scLDA}
\title{Runs scLDA}
\usage{
scLDA(
  data,
  parameter,
  nIter = 20,
  error_out_threshold = 1e-04,
  nIter_inner = 20,
  error_inner_threshold = 1e-04
)
}
\arguments{
\item{data}{the input data list. The input
data is a list of two datasets, scRNAseq and bulk RNAseq.}

\item{parameter}{the vector of parameters. The first parameter is the value of alpha in the mathematical model
, the second one is the value of beta in the mathematical model.}

\item{nIter}{the maximum iterations, the default is 20.}

\item{error_out_threshold}{the threshold of the error between the current imputed matrix and the previous one.
Default is 1e-4.}

\item{nIter_inner}{the maximum interations of calculating the sub-optimization problem. Default is 20.}

\item{error_inner_threshold}{the threshold of the error between the current updated matrix and the previous one.
Default is 1e-4.}
}
\value{
A data matrix with the same size of the input scRNAseq data
}
\description{
MCEM version of mtSCRABBLE (multi-task SCRABBLE) using LDA to get the deconvolution weight for A
}
\examples{
# Set up the parameter used in SCRABBLE
parameter <- c(1, 1e-6, 1e-4)

# Run SCRABLE
result <- scrabble(demo_data,parameter = parameter)

}
