% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score matrix columns for groups of rows}
\usage{
score(mat, groups, controls = NULL, center = F, bin.control = F,
  bins = NULL, binmat = NULL, nbin = 30, replace = F,
  bin.size = 100)
}
\arguments{
\item{mat}{a matrix.}

\item{groups}{a character vector or list of character vectors. Each character vector is a group or signature to score each column against.}

\item{controls}{PA Default: NULL}

\item{center}{boolean value. Should the resulting score matrix be row-centered? Default: F}

\item{bin.control}{boolean value. Should control groups be generated by binning the rows? Default: F}

\item{bins}{a named character vector with as names the rownames and as values the ID of each bin. Default: NULL}

\item{binmat}{a matrix. If control groups are to be generated from row bins, the user can provide a different matrix of values by which to bin the rownames. The rownames should then obviously match those in <mat>. If NULL, the matrix provided in <mat> will be used. Default: NULL}

\item{nbin}{numeric value. How many bins should be created? Default: 30}

\item{replace}{boolean value. Allow bin sampling to be done with replacement. Default: F}

\item{bin.size}{numeric value. How big should the bins be? Default: 100}
}
\value{
a matrix with as rows the columns of the input matrix and as columns the scores of each group provided in groups. If one group is provided, the matrix returned will have 1 column.
}
\description{
Score a matrix by groups of rows. Scores are column averages for the subset of rows specified in a group. Option to generate control group scores to subtract from the real scores. Control groups can either be user-provided or generated from binning of the rows. Similarly, the bins themselves can be user-provided or computed.
}
\details{
DETAILS
}
