% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{t_tests}
\alias{t_tests}
\title{t_tests}
\usage{
t_tests(m, m2 = NULL, group = NULL, adjust.method = "BH",
  MARGIN = 1)
}
\arguments{
\item{m}{matrix 1}

\item{m2}{matrix 2. If null, m2 = m[, !group]. Defaut: NULL}

\item{group}{if m2 not supplied, m = m[, group]; m2 = m[, !group]. Default: NULL}

\item{adjust.method}{Correction for multiple tests. If not desired, set to 'none'. Default: 'BH'}

\item{MARGIN}{matrix margin to apply t-test by. If set to 2, T-test instead between pairs of columns. Default: 1}
}
\value{
numeric vector of (adjusted) p-values
}
\description{
T-test between pairs of rows from two matrices
}
\details{
DETAILS
}
\seealso{
\code{\link[stats]{t.test}},\code{\link[stats]{p.adjust}}
}
