% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquely.R
\name{uniquely}
\alias{uniquely}
\title{uniquely}
\usage{
uniquely(L, Names = F, highest = T)
}
\arguments{
\item{L}{List of named numeric vectors (e.g. Gene names and FC values for different clusters)}

\item{Names}{return just names (Genes), Default: F}

\item{highest}{if TRUE, unique instance is equal to highest instance, else to lowest. Default: T}
}
\value{
List of named numeric vectors, with only one instance of each name across all vectors
}
\description{
unique instances of names (Genes) across a list of named numeric vectors
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}}
}
