% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_coord.R
\name{build_umap}
\alias{build_umap}
\title{Uniform Manifold Approximation and Projection}
\usage{
build_umap(mat, dist.method = "pearson", ...)
}
\arguments{
\item{mat}{input matrix (each row is a variable / a gene and each column is an observation / a cell.}

\item{dist.method}{a character string for the distance metric used. One of pearson, euclidean, manhattan, cosine or pearson2. Default: 'pearson'}

\item{...}{other arguments passed to umap}
}
\value{
coordinates embedding matrix
}
\description{
Computes a manifold approximation and projection and returns the embedding coordinates matrix. Wrapper function for umap::umap().
}
\seealso{
umap, umap.defaults
}
