% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_by_size.R
\name{clean_by_size}
\alias{clean_by_size}
\title{clean_by_size}
\usage{
clean_by_size(clusters, min.rel = 0.01, max.rel = 0.8, min.abs = 5)
}
\arguments{
\item{clusters}{List of cluster memberships (character vectors)}

\item{min.rel}{minimum relative size of a cluster. Default: 0.01 (ie. 0.01 of total number of variables)}

\item{max.rel}{maximum relative size of a cluster. Default: 0.8 (ie. 0.8 of total number of variables)}

\item{min.abs}{absolute minimum size. If min.rel < min.abs, min.abs will replace min.rel. Default: 5}
}
\value{
a list of filtered clusters according to their being within min.rel and max.rel cutoffs.
}
\description{
Filtering for Reasonably Sized Clusters
}
\details{
DETAILS
}
