% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA}
\alias{DEA}
\title{Differential Gene Expression Analysis across Cell Clusters}
\usage{
DEA(m, clusters, FC = 2, is.log = T, p = 0.01,
  adjust.method = "BH", sort = T, p.sort = F, return.full = F,
  return.p = F, fast = T)
}
\arguments{
\item{m}{matrix}

\item{clusters}{List of cluster memberships (character vectors). If provided, will be used instead of computed.. Default: NULL}

\item{FC}{fold change cutoff. Default: 2}

\item{is.log}{values are in log2 (used in fold_changes). Default: T}

\item{p}{p-value cutoff. Default: 0.01}

\item{adjust.method}{Correction for multiple tests. If not desired, set to 'none'. Default: 'BH'}

\item{sort}{returning values are sorted by fold change (largest first). Default: T}

\item{p.sort}{returning values are sorted by p-value (most significant first). Default: F}

\item{return.full}{return cluster membership and cluster DE genes. Default: F}

\item{return.p}{return cluster DE p-values. Default: F}

\item{fast}{if TRUE, computes p-values for significant fold-change values only. Default: T}
}
\value{
with default parameters, will return a list of differentially expressed genes in statistically significant clusters. Differentially expressed genes are those passing fold change and p-value cutoffs. Statistically significant clusters are those passing nsig1 and nsig2 cutoffs. Within clusters, gene values displayed are fold changes and genes are ordered by highest FC first.
}
\description{
This function returns the genes differentially expressed in a given cluster of cells relative to the remaining cells in the matrix. If a list of clusters is provided, DEA will be computed for each cluster in turn. Differential expression is calculated according to fold changes and p-values. The user can specify the fold change and p-value with which genes are deemed (DE-)significant. With default arguments, genes are considered significant if FC >= 2 and p-value
(Benjamin-Hochberg-adjusted) <= 0.01. Also with default arguments, fold-change values are returned and genes are sorted by fold-change (largest first). The user can instead ask for genes to be sorted by p-value (most significant first) or for p-values to be returned or both. Alternatively, if the user requires fold change values and p-values for all genes (not just those that are significant), they can set return.full = TRUE. This parameter will return a list containing (1) all clusters, (2) fold-change values for all genes, (3) p-values for all genes, (4) the gene names and indices of the DE-significant genes and (5-7) the fold change and p-value cutoffs used as well as the method used to adjust p-values for multiple testing.
}
\details{
DETAILS
}
