% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea}
\alias{dea}
\title{Differential Expression Analysis}
\usage{
dea(x, y, is.log = T, cutoff.fc = 2, cutoff.p = 0.01,
  adjust.method = unique(c("BH", stats::p.adjust.methods)),
  sortBy = c("fc", "p", "none"), returnValue = c("fc", "p", "all"),
  returnAllGenes = FALSE)
}
\arguments{
\item{x}{matrix or character vector with matrix column names}

\item{y}{matrix to compare against or matrix with x and y column names}

\item{is.log}{is the expression data in log2 space? Default: T}

\item{cutoff.fc}{numeric value for fold change gene cutoffs (not in log space) or NULL. Default: 2}

\item{cutoff.p}{numeric value for p-value gene cutoffs (not in log space) or NULL. Default: 0.01}

\item{adjust.method}{correction for multiple testing? 'none' if not desired. Default: unique(c("BH", stats::p.adjust.methods))}

\item{sortBy}{parameter to sort the genes returned by. One of fold change (default), p-value or none, in which case the original order of the rows in the matrix is kept. Default: c("fc", "p", "none")}

\item{returnValue}{parameter describing the genes to be returned. One of fold change (default), p-value or all,  in which case fold changes, p-values and genes considered 'DE' (ie if cutoffs are provided) are returned.Default: c("fc", "p", "all")}

\item{returnAllGenes}{should all genes be returned instead of only those passing cutoffs? Default: FALSE}
}
\value{
if returnValue == 'all', a list, else a named numeric vector. If a named numeric vector, names are gene names (row names) and values are fold changes if returnValue == 'fc' or p values of returnValue == 'p'. The number of genes returned depends on whether returnAllGenes == T and whether cutoff.fc and cutoff.p were given (default is that they are).
}
\description{
differential expression analysis
}
\seealso{
\code{\link[stats]{character(0)}}
}
